fix(restapi): validate pagination bounds and answer them in JSON
Out-of-range ?page was silently wrapped and served the wrong rows: the spec
typed page as uint32 while handlers narrowed to uint16, so ?page=65537 came
back as page 1 with a 200, and /requests computed its offset in wrapping
uint32 arithmetic — page=42949674&limit=100 returned rows 5..104. page is now
uint16 with its bound declared, so an out-of-range value fails the bind.

The endpoints also disagreed with each other at both ends. page=0 was a 500 on
/movies and silent coercion elsewhere; limit=0 answered 200 with a limit field
claiming 0 while serving 20; a limit above the documented maximum was silently
clamped on some endpoints and rejected at bind on others; and /activity,
/activity/history and /users passed a zero straight to the store, which
substituted its own default. Every list endpoint now answers a present page or
limit outside its documented range — zero, negative, or over the maximum —
with a 400 naming the bound, so the spec's minimum/maximum are an enforced
contract rather than documentation.

Those 400s are JSON. Binding failures previously fell through to the
generated text/plain http.Error on an API where every other response is JSON,
because Mount never set ChiServerOptions.ErrorHandlerFunc; both it and the
strict handler's request-error path now emit the standard error shape, with
the parameter's name rather than the Go type it failed to fit.

Counts move to int end-to-end underneath this. Every list total and status
count originates as a Go int — ent's Count() or a len() — and was narrowed at
~25 sites only because the spec declared x-go-type there, each needing a G115
suppression, several of which golines had silently broken by wrapping the
line. The JSON wire format is unchanged; ent's storage columns stay narrow and
convert at one choke point each. schedules' last_duration_ms is aligned with
its uint32 storage the same way, keeping portable bounds so non-Go generators
still see a range.

Library listings are capped at 500 rather than 100: the SPA fetches movies,
series and requests whole and filters client-side, so the dashboard tiles, the
search index and the "already added" maps need the full list.

+++ internal/library/bulkimport/bulkimport_test.go
@@ -133,7 +133,7 @@ var _ = Describe("Service file decisions", Label("unit", "bulkimport"), func() {
 						return p.ScanID == 3
 					},
 				)).
-				Return([]*ent.ImportScanFile{}, uint32(0), nil).
+				Return([]*ent.ImportScanFile{}, 0, nil).
 				Once()
 
 			items, total, err := svc.Files(ctx, FilesParams{ScanID: 3, Limit: 50})
